refactor: accents (cont)
