+++ src/server/domain/chantier/ChantierRepository.interface.ts
@@ -11,7 +11,7 @@ import { AvancementsStatistiques } from "@/components/_commons/Avancements/Avanc
 import { ProfilCode } from "@/server/domain/utilisateur/Utilisateur.interface";
 import { RepartitionMeteoChantiers } from "@/server/chantiers/domain/RepartitionMeteoChantiers";
 import { FiltreQueryParams } from "@/server/chantiers/app/contrats/FiltreQueryParams";
-import { ChantierPourAgregation } from "@/client/utils/chantier/agrégateurListeChantiers/agrégateur";
+import { ChantierPourAgregation } from "@/client/utils/chantier/agrégateurListeChantiers/agregateur";
 
 export default interface ChantierRepository {
   récupérerLesEntréesDUnChantier(
